[COLAB-2918]: change goal to 'generatePojos' and parameter 'packageSuffix' to 'implementationPackageSuffix'; refactor ISubLocation in test-directory; implement codacy suggestions
