+++ other/pojo-generator-maven-plugin/pom.xml
@@ -9,7 +9,6 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <groupId>org.xcolab</groupId>
     <artifactId>pojo-generator-maven-plugin</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>maven-plugin</packaging>
