+++ other/pojo-generator-maven-plugin/src/test/resources/project-to-test/org/xcolab/pojo/generator/ILocation.java
@@ -4,10 +4,11 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
 import org.xcolab.pojo.generator.tables.pojos.Location;
 
+import java.io.Serializable;
 import java.sql.Timestamp;
 
 @JsonDeserialize(as = Location.class)
-public interface ILocation {
+public interface ILocation extends Serializable {
 
     String getCity();
